// components/client-disp-dev/client-disp-dev/src/pages/work-time-map/lib/types/timeline-range-changed-event.ts
/**
 * Представляет событие получаемое при окончании изменения диапазона линии времени.
 */
export interface TimelineRangeChangedEvent {
  /** Возвращает дату начала видимой области линии времени. */
  readonly start: Date;
  /** Возвращает дату концы видимой области линии времени. */
  readonly end: Date;
}